+++ FoundationDB.Client/Layers/Tuples/FdbTupleTypes.cs
@@ -36,50 +36,50 @@ namespace FoundationDB.Layers.Tuples
 	internal static class FdbTupleTypes
 	{
 		/// <summary>Null/Empty/Void</summary>
-		internal const byte Nil = (byte)0;
+		internal const byte Nil = 0;
 
 		/// <summary>ASCII String</summary>
-		internal const byte Bytes = (byte)1;
+		internal const byte Bytes = 1;
 
 		/// <summary>UTF-8 String</summary>
-		internal const byte Utf8 = (byte)2;
+		internal const byte Utf8 = 2;
 
 		/// <summary>Nested tuple [DRAFT]</summary>
-		internal const byte TupleStart = (byte)3;
+		internal const byte TupleStart = 3;
 		/// <summary>End of a nested tuple [DRAFT]</summary>
-		internal const byte TupleEnd = (byte)4;
+		internal const byte TupleEnd = 4;
 
-		internal const byte IntNeg8 = (byte)12;
-		internal const byte IntNeg7 = (byte)13;
-		internal const byte IntNeg6 = (byte)14;
-		internal const byte IntNeg5 = (byte)15;
-		internal const byte IntNeg4 = (byte)16;
-		internal const byte IntNeg3 = (byte)17;
-		internal const byte IntNeg2 = (byte)18;
-		internal const byte IntNeg1 = (byte)19;
-		internal const byte IntZero = (byte)20;
-		internal const byte IntPos1 = (byte)21;
-		internal const byte IntPos2 = (byte)22;
-		internal const byte IntPos3 = (byte)23;
-		internal const byte IntPos4 = (byte)24;
-		internal const byte IntPos5 = (byte)25;
-		internal const byte IntPos6 = (byte)26;
-		internal const byte IntPos7 = (byte)27;
-		internal const byte IntPos8 = (byte)28;
+		internal const byte IntNeg8 = 12;
+		internal const byte IntNeg7 = 13;
+		internal const byte IntNeg6 = 14;
+		internal const byte IntNeg5 = 15;
+		internal const byte IntNeg4 = 16;
+		internal const byte IntNeg3 = 17;
+		internal const byte IntNeg2 = 18;
+		internal const byte IntNeg1 = 19;
+		internal const byte IntZero = 20;
+		internal const byte IntPos1 = 21;
+		internal const byte IntPos2 = 22;
+		internal const byte IntPos3 = 23;
+		internal const byte IntPos4 = 24;
+		internal const byte IntPos5 = 25;
+		internal const byte IntPos6 = 26;
+		internal const byte IntPos7 = 27;
+		internal const byte IntPos8 = 28;
 
 		/// <summary>Base value for integer types (20 +/- n)</summary>
 		internal const int IntBase = 20;
 
 		/// <summary>RFC4122 UUID (128 bits) [DRAFT]</summary>
-		internal const byte Guid = (byte)48;
+		internal const byte Guid = 48;
 
 		/// <summary>Standard prefix of the Directory Layer</summary>
 		/// <remarks>This is not a part of the tuple encoding itself, but helps the tuple decoder pretty-print tuples that would otherwise be unparsable.</remarks>
-		internal const int AliasDirectory = (byte)254;
+		internal const byte AliasDirectory = 254;
 
 		/// <summary>Standard prefix of the System keys, or frequent suffix with key ranges</summary>
 		/// <remarks>This is not a part of the tuple encoding itself, but helps the tuple decoder pretty-print End keys from ranges, that would otherwise be unparsable.</remarks>
-		internal const int AliasSystem = (byte)255;
+		internal const byte AliasSystem = 255;
 	}
 
 }
